new global reco with bbc,epd and globalvertexmap

+++ submit/fm_0_20/pass3global/rundir/run_pass3global.sh
@@ -11,17 +11,20 @@ this_dir=`dirname $this_script`
 echo rsyncing from $this_dir
 echo running: $this_script $*
 
-source /cvmfs/sphenix.sdcc.bnl.gov/gcc-12.1.0/opt/sphenix/core/bin/sphenix_setup.sh -n ana.349
+source /cvmfs/sphenix.sdcc.bnl.gov/gcc-12.1.0/opt/sphenix/core/bin/sphenix_setup.sh -n ana.354
 
 
 if [[ ! -z "$_CONDOR_SCRATCH_DIR" && -d $_CONDOR_SCRATCH_DIR ]]
 then
     cd $_CONDOR_SCRATCH_DIR
     rsync -av $this_dir/* .
-    getinputfiles.pl  $2
+    echo $2 > inputfiles.list
+    echo $3 >> inputfiles.list
+    getinputfiles.pl --filelist inputfiles.list
     if [ $? -ne 0 ]
     then
-	echo error from getinputfiles.pl $2, exiting
+        cat inputfiles.list
+	echo error from getinputfiles.pl  --filelist inputfiles.list, exiting
 	exit -1
     fi
 else
@@ -31,37 +34,36 @@ fi
 # arguments 
 # $1: number of events
 # $2: bbc g4hits input file
-# $3: output file
-# $4: output dir
-# $5: run number
-# $6: sequence
+# $3: truth g4hits input file
+# $4: output file
+# $5: output dir
+# $6: run number
+# $7: sequence
 
 echo 'here comes your environment'
 printenv
 echo arg1 \(events\) : $1
 echo arg2 \(bbc g4hits file\): $2
-echo arg3 \(output file\): $3
-echo arg4 \(output dir\): $4
-echo arg5 \(runnumber\): $5
-echo arg6 \(sequence\): $6
+echo arg3 \(truth g4hits file\): $3
+echo arg4 \(output file\): $4
+echo arg5 \(output dir\): $5
+echo arg6 \(runnumber\): $6
+echo arg7 \(sequence\): $7
 
-runnumber=$(printf "%010d" $5)
-sequence=$(printf "%05d" $6)
+runnumber=$(printf "%010d" $6)
+sequence=$(printf "%05d" $7)
 filename=fm_0_20_pass3global
 
 txtfilename=${filename}-${runnumber}-${sequence}.txt
 jsonfilename=${filename}-${runnumber}-${sequence}.json
 
-echo running prmon  --filename $txtfilename --json-summary $jsonfilename -- root.exe -q -b Fun4All_G4_Global.C\($1,\"$2\",\"$3\",\"$4\"\)
-prmon  --filename $txtfilename --json-summary $jsonfilename -- root.exe -q -b  Fun4All_G4_Global.C\($1,\"$2\",\"$3\",\"$4\"\)
+echo running prmon  --filename $txtfilename --json-summary $jsonfilename -- root.exe -q -b Fun4All_G4_Global.C\($1,\"$2\",\"$3\",\"$4\",\"$5\"\)
+prmon  --filename $txtfilename --json-summary $jsonfilename -- root.exe -q -b  Fun4All_G4_Global.C\($1,\"$2\",\"$3\",\"$4\",\"$5\"\)
 
 rsyncdirname=/sphenix/user/sphnxpro/prmon/fm_0_20/pass3global/${runnumber}
-if [ ! -d $rsyncdirname ]
-then
-  mkdir -p $rsyncdirname
-fi
+[ -d $rsyncdirname ] || mkdir -p $rsyncdirname
 
-rsync -av $txtfilename $rsyncdirname
-rsync -av $jsonfilename $rsyncdirname
+[ -f $txtfilename ] && rsync -av $txtfilename $rsyncdirname
+[ -f $jsonfilename ] && rsync -av $jsonfilename $rsyncdirname
 
 echo "script done"
